# src/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API MemoBelc",
    "description": "Documentação da API MemoBelc. Use **Authorize** para informar o token Bearer e testar rotas protegidas."
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token JWT: use o valor 'Bearer <seu_token>'"
    }
  },
  "security": [],
...
  "paths": {
...
    "/deck/create": {
      "post": {
        "tags": ["Decks"],
        "summary": "Criar baralho",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "collection_id"],
              "properties": {
                "name": { "type": "string" },
                "collection_id": { "type": "string" },
                "image": { "type": "string" },
                "cards": { "type": "array" }
              }
            }
          }
        ],
        "responses": { "200": { "description": "Baralho criado" }, "400": { "description": "Dados obrigatórios" } }
      }
    },
...
    "/deck/get_by_collection_id": {
      "get": {
        "tags": ["Decks"],
        "summary": "Baralhos por coleção",
        "parameters": [
          { "name": "collection_id", "in": "query", "required": true, "type": "string" },
          { "name": "user_id", "in": "query", "required": true, "type": "string" }
        ],
        "responses": { "200": { "description": "Lista de baralhos" }, "400": { "description": "Parâmetros obrigatórios" } }
      }
    },
    "/deck/save_deck": {
      "post": {
        "tags": ["Decks"],
        "summary": "Salvar baralho na coleção do usuário",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["user_id", "deck_id", "collection_id"],
              "properties": {
                "user_id": { "type": "string" },
                "deck_id": { "type": "string" },
                "collection_id": { "type": "string" }
              }
            }
          }
        ],
        "responses": { "200": { "description": "Salvo" }, "400": { "description": "Dados obrigatórios" } }
      }
    },
    "/deck/check_if_the_user_has_the_deck": {
      "post": {
        "tags": ["Decks"],
        "summary": "Verificar se usuário possui o baralho",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["user_id", "deck_id"],
              "properties": { "user_id": { "type": "string" }, "deck_id": { "type": "string" } }
            }
          }
        ],
        "responses": { "200": { "description": "Resposta com flag" }, "400": { "description": "Dados obrigatórios" } }
      }
    },
...
  }
}